Moved the project files to the hw2 directory, and implemented the function of generating 100 random names and addresses
