+++ hw2/requirements.txt
@@ -1,0 +1,5 @@
+pip~=21.1.3
+python-dateutil~=2.8.1
+six~=1.16.0
+setuptools~=57.0.0
+Flask~=2.0.1
