--- requirements.txt
@@ -1,5 +1,0 @@
-pip~=21.0.1
-python-dateutil~=2.8.1
-six~=1.15.0
-setuptools~=49.2.1
-Flask~=1.1.2
